+++ src/index.ts
@@ -1,25 +1,46 @@
 import css from './css';
 import buildStyles from './css/render';
-import createStyler from './styler';
-import { Styler, Props } from './styler/types';
 import svg from './svg';
+import viewport from './viewport';
+import createStylerFactory from './styler';
+import { invariant } from 'hey-listen';
+import { Styler, Props } from './styler/types';
 
-const cache = new WeakMap<Element, Styler>();
+const cache = new WeakMap<Element | Window, Styler>();
 
-const createDOMStyler = (node: Element, props: Props) => {
-  const styler = (node instanceof SVGElement) ? svg(node) : css(node as HTMLElement, props);
+const createDOMStyler = (node: Element | Window, props: Props) => {
+  let styler: Styler;
+
+  if (node instanceof HTMLElement) {
+    styler = css(node, props);
+  } else if (node instanceof SVGElement) {
+    styler = svg(node);
+  } else if (typeof window !== 'undefined' && node === window) {
+    styler = viewport(node);
+  }
+
+  invariant(
+    styler !== undefined,
+    'No valid node provided. Node must be HTMLElement, SVGElement or window.'
+  );
+
   cache.set(node, styler);
   return styler;
 };
 
-const getStyler = (node: Element, props: Props) => cache.has(node) ? cache.get(node) : createDOMStyler(node, props);
+const getStyler = (node: Element | Window, props: Props) =>
+  cache.has(node) ? cache.get(node) : createDOMStyler(node, props);
 
-export default function(nodeOrSelector: Element | string, props: Props): Styler {
-  const node: Element = (typeof nodeOrSelector === 'string')
-    ? document.querySelector(nodeOrSelector)
-    : nodeOrSelector;
+export default function(
+  nodeOrSelector: Element | string | Window,
+  props: Props
+): Styler {
+  const node: Element | Window =
+    typeof nodeOrSelector === 'string'
+      ? document.querySelector(nodeOrSelector)
+      : nodeOrSelector;
 
   return getStyler(node, props);
 }
 
-export { createStyler, Styler, buildStyles };
+export { createStylerFactory, Styler, buildStyles };